even store design - need to update the procedure for the Event store. Dispatching. When is the proper case for publishing the events?

+++ EventSourcing/HP.Application/Commands/DeleteTodoCommandHandler.cs
@@ -14,13 +14,15 @@ namespace HP.Application.Commands
         }
         public async Task<bool> Handle(DeleteTodoCommand request, CancellationToken cancellationToken)
         {
+            var todo = await _repository.GetByIdAsync(request.Id);
+            if(todo == null)
+                throw new ArgumentNullException(nameof(todo));
 
-
+            todo.DeleteTodo
             Expression<Func<Todo, bool>> expr = x => x.Id == request.Id;
             await _repository.DeleteOneAsync(expr);
             var @event = new TodoDomainEvents.TodoRemovedEvent(request.Id);
             // Publish Remove event. 
-
             return true;
         }
     }
